+++ tools/build-xilinx-vivado.rb
@@ -1,5 +1,10 @@
 #!/usr/bin/env ruby
 
+puts "***** This script is deprecated and no longer maintained *****"
+puts ""
+puts "Use `nvc --install vivado' instead."
+exit 1    # Comment this if you really want to run it
+
 #
 # Script to compile the Xilinx Vivado simulation libraries
 #
@@ -37,7 +42,7 @@ FileUtils.mkdir_p $libdir
 
 def run_nvc(lib, file)
   file = "#{$src}/#{file}" unless file =~ /^\//
-  cmd = "nvc --work=#{$libdir}/#{lib} -a --relaxed #{file}"
+  cmd = "nvc -M 32m --work=#{$libdir}/#{lib} -a --relaxed #{file}"
   puts cmd
   exit 1 unless system cmd
 end